backend/requirements.txt
fastapi
uvicorn[standard]
httpx
aiosqlite
python-dotenv
pydantic
